+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+# KaliberX Changelog
+
+## 1.0.0 - TBD
+
+Initial release
